chore: reapply #476 on new dashboard
