+++ src/extension/options-page/DashboardComponents/FooterLine.tsx
@@ -1,0 +1,54 @@
+import React from 'react'
+import { makeStyles, createStyles } from '@material-ui/styles'
+import { Link } from 'react-router-dom'
+import { Breadcrumbs, Theme, Typography, Link as MuiLink } from '@material-ui/core'
+import { geti18nString } from '../../../utils/i18n'
+
+const useStyles = makeStyles((theme: Theme) =>
+    createStyles({
+        footerButtons: {
+            marginTop: theme.spacing(1),
+            '& ol': {
+                justifyContent: 'center',
+            },
+        },
+        footerButton: {
+            borderRadius: '0',
+            whiteSpace: 'nowrap',
+            '& > span': {
+                marginLeft: theme.spacing(1),
+                marginRight: theme.spacing(1),
+            },
+        },
+    }),
+)
+
+export default function FooterLine() {
+    const classes = useStyles()
+    const FooterLink = function(props: any) {
+        return (
+            <MuiLink
+                underline="none"
+                {...(props.href
+                    ? { href: props.href, target: '_blank', rel: 'noopener noreferrer' }
+                    : { to: props.to, component: Link })}
+                color="inherit"
+                className={classes.footerButton}>
+                <Typography variant="body2">{props.children}</Typography>
+            </MuiLink>
+        )
+    }
+    return (
+        <Breadcrumbs className={classes.footerButtons} separator="|" aria-label="breadcrumb">
+            <FooterLink href="https://maskbook.com/">Maskbook.com</FooterLink>
+            <FooterLink href="https://github.com/DimensionDev/Maskbook/releases">Version 1.8.0</FooterLink>
+            <FooterLink href="https://maskbook.com/download-links/#mobile">
+                {geti18nString('dashboard_mobile_test')}
+            </FooterLink>
+            <FooterLink href="https://github.com/DimensionDev/Maskbook">
+                {geti18nString('dashboard_source_code')}
+            </FooterLink>
+            <FooterLink href="https://maskbook.com/privacy-policy/">Privacy Policy</FooterLink>
+        </Breadcrumbs>
+    )
+}
